+++ test/Spec.hs
@@ -7,6 +7,7 @@ import Data.SRTree.Eval (compile)
 import Algorithm.SRTree.AD.Unboxed (CompiledTree, compileTree, compileTreeMulti, evalGrad, evalGradVec, evalGradMulti)
 import qualified EqSatTests
 import qualified StoreTests
+import qualified CITests
 import Data.SRTree.Random (randomTree, tossBiased, randomFrom)
 import System.Random (mkStdGen)
 import Control.Monad.State.Strict (evalStateT)
@@ -109,6 +110,7 @@ main = do
     , TestLabel "benchgrad" test_benchgrad
     , TestLabel "eqsat" EqSatTests.tests
     , TestLabel "store" StoreTests.tests
+    , CITests.tests
     ]
   if failures counts /= 0 || errors counts /= 0
     then error "Some tests failed"